add jabs-classify script as part of package
